:recycle: refactor: 컨텐츠 상세 페이지 컴포넌트 구조 분리
