+++ src/widgets/content/feature/useHandleLike.ts
@@ -1,0 +1,13 @@
+import { useState } from "react";
+
+export const useHandleLike = () => {
+  const [isLiked, setIsLiked] = useState(false);
+  const [likeCount, setLikeCount] = useState(124);
+
+  const handleLike = () => {
+    setIsLiked(!isLiked);
+    setLikeCount((prev) => (isLiked ? prev - 1 : prev + 1));
+  };
+
+  return { isLiked, likeCount, handleLike };
+};
